+++ anta/_advisory/reporter/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2026 Arista Networks, Inc.
+# Use of this source code is governed by the Apache License 2.0
+# that can be found in the LICENSE file.
+"""Internal reporting support for ANTA security advisory results."""
